requirements.txt
langchain
langchain-ibm
langchain-community
ibm-watsonx-ai
ibm-watson-machine-learning
tiktoken
python-dotenv
streamlit
pandas
fpdf